feat: add code quality workflows
This adds a workflow ci, and various code quality checks.
It also adds a shellscript with the same checks so
that the developer has a chance of running a check
on his feature branch BEFORE the pipeline checks.

Note that the license (reuse) check will fail
currently, and that is on purpose - that will be
adressed in next pr, seperatly, due to it adding a header in many files.

Co-authored-by: Johannes Dannelöv <johannes.dannelov@digg.se>
Co-authored-by: Konstantin Ay <konstantin.ay@intunio.se>
Signed-off-by: Josef Andersson <josef.andersson@digg.se>